use bonny to apply descendants
