+++ spawn_operator/spawn_operator/lib/spawn_operator/handler/actor_host_handler.ex
@@ -1,13 +1,32 @@
 defmodule SpawnOperator.Handler.ActorHostHandler do
   @moduledoc """
   `ActorHostHandler` handles ActorHost CRD events
+
+      ---
+      apiVersion: spawn-eigr.io/v1
+      kind: ActorHost
+      metadata:
+        name: my-node-app # Mandatory. Name of the Node containing Actor Host Functions
+        system: my-actor-system # mandatory. Name of the ActorSystem declared in ActorSystem CRD
+        namespace: default # Optional. Default namespace is "default"
+      spec:
+        affinity: k8s_affinity_declaration_here # Optional
+
+        replicas: 1 # Optional. If negative number than autoscaling is enable
+
+        host: # Mandatory
+          image: docker.io/eigr/spawn-springboot-examples:latest # Mandatory
+          embedded: false # Optional. Default false. True only when the SDK supports a native connection to the Spawn mesh network
+          ports:
+          - containerPort: 80
+
+        sidecar: # Optional. If embedded true then this section will be ignored
+          image: docker.io/eigr/spawn-proxy:0.1.0
+
   """
 
-  import SpawnOperator,
-    only: [
-      build_actor_host: 1,
-      track_event: 2
-    ]
+  alias SpawnOperator.K8s.ConfigMap.SidecarCM
+  alias SpawnOperator.K8s.Deployment
 
   @behaviour Pluggable
 
@@ -16,14 +35,29 @@ defmodule SpawnOperator.Handler.ActorHostHandler do
 
   @impl Pluggable
   def call(%Bonny.Axn{action: action} = axn, nil) when action in[:add, :modify] do
-    track_event(action, axn.resource)
-    build_actor_host(axn.resource)
-    Bonny.Axn.success_event(axn)
+    %Bonny.Axn{resource: resource} = axn
+    host_resource = build_host_deploy(resource)
+    host_config_map = build_host_configmap(resource)
+
+    axn
+    |> Bonny.Axn.register_descendant(host_resource)
+    |> Bonny.Axn.register_descendant(host_config_map)
+    |> Bonny.Axn.success_event()
   end
 
   @impl Pluggable
   def call(%Bonny.Axn{action: action} = axn, nil) when action in[:delete, :reconcile] do
-    track_event(action, axn.resource)
     Bonny.Axn.success_event(axn)
   end
+
+  defp build_host_deploy(resource) do
+    %{system: system, namespace: ns, name: name, params: params} = SpawnOperator.get_args(resource)
+    Deployment.manifest(system, ns, name, params)
+  end
+
+  defp build_host_configmap(resource) do
+    %{system: system, namespace: ns, name: name, params: params} = SpawnOperator.get_args(resource)
+    SidecarCM.manifest(system, ns, name, params)
+  end
+
 end
